+++ src/main/java/com/proyecto/medihealth/administrador/repositories/HistoriaClinicaRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface HistoriaClinicaRepository extends JpaRepository<HistoriaClinica, Integer> {
 
-    Optional<HistoriaClinica> findByNumeroHistoria(long numeroHistoria);
+    Optional<HistoriaClinica> findByNumeroHistoria(int numeroHistoria);
     Optional<HistoriaClinica> findByPaciente_Usuario_DocumentoIdentidad(String documentoIdentidad);
